updated db sql & queries sql done

+++ SQL/DataBase/Script-1.sql
@@ -32,7 +32,7 @@ CREATE TABLE gourmetdetallespedidos (
 id_pedido INT NOT NULL,
 id_menu INT NOT NULL,
 cantidad INT NOT NULL,
-precio_unitario INT NOT NULL, 
+precio_unitario DECIMAL(10,2) NOT NULL, 
 PRIMARY KEY (id_pedido, id_menu),
 FOREIGN KEY (id_pedido) REFERENCES gourmetpedidos(id_pedido),
 FOREIGN KEY (id_menu) REFERENCES gourmetmenus(id_menu)
